LNK-1246: QA needs conditional API endpoints for DELETE functionality (#569)
* Delete `MeasureReport`s before `Report`s

* Improve reporting for deletion failure

+++ api/src/main/java/com/lantanagroup/link/api/controller/QAController.java
@@ -3,19 +3,24 @@ package com.lantanagroup.link.api.controller;
 import com.lantanagroup.link.config.api.ApiConfig;
 import com.lantanagroup.link.db.SharedService;
 import com.lantanagroup.link.db.TenantService;
+import com.lantanagroup.link.db.model.Report;
+import com.lantanagroup.link.model.ReportBase;
 import lombok.Setter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.ApplicationContext;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/qa/{tenantId}")
@@ -51,9 +56,16 @@ public class QAController extends BaseController {
   }
 
 
-  @DeleteMapping("/$deleteByPatientListId/{id}")
+  @DeleteMapping("/$deletePatientListById/{id}")
   public void deletePatientListById(@PathVariable String tenantId, @PathVariable UUID id){
     TenantService tenantService = TenantService.create(this.sharedService, tenantId);
+    List<Report> reports = tenantService.getReportsByPatientListId(id);
+    if (!reports.isEmpty()) {
+      String message = String.format(
+              "The following reports reference the specified patient list: %s",
+              reports.stream().map(ReportBase::getId).collect(Collectors.joining(", ")));
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST, message);
+    }
     tenantService.deletePatientListById(id);
   }
 
@@ -69,7 +81,7 @@ public class QAController extends BaseController {
     tenantService.deletePatientByListAndPatientId(patientId,patientListId);
   }
 
-  @DeleteMapping("/$deletebyReportId/{reportId}")
+  @DeleteMapping("/$deleteReportById/{reportId}")
   public void deleteReport(@PathVariable String tenantId,@PathVariable String reportId){
     TenantService tenantService = TenantService.create(this.sharedService, tenantId);
     tenantService.deleteReport(reportId);
